+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+## [Unreleased]
+
 ## [1.0.0-beta.2]
 ### Added
 - Added support to manually set bulk size and enable/disable bulk dispatch