+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [1.0.0-beta.3]
 ### Added
 - Ability to use timeOffset only when local system is using non-UTC timezone
 - Support to replace existing logging libraries with ease